+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/jyj/Documents/Projects/showmemydog/src/index.tsx":"1","/Users/jyj/Documents/Projects/showmemydog/src/Routes.tsx":"2","/Users/jyj/Documents/Projects/showmemydog/src/Pages/Quiz/index.tsx":"3","/Users/jyj/Documents/Projects/showmemydog/src/Pages/Result/index.tsx":"4","/Users/jyj/Documents/Projects/showmemydog/src/Pages/Main/index.tsx":"5","/Users/jyj/Documents/Projects/showmemydog/src/Styles/globalStyles.ts":"6","/Users/jyj/Documents/Projects/showmemydog/src/Styles/theme.ts":"7","/Users/jyj/Documents/Projects/showmemydog/src/API.ts":"8","/Users/jyj/Documents/Projects/showmemydog/src/Pages/Quiz/cards.tsx":"9"},{"size":246,"mtime":1611927354408,"results":"10","hashOfConfig":"11"},{"size":593,"mtime":1611927354407,"results":"12","hashOfConfig":"11"},{"size":929,"mtime":1612086745516,"results":"13","hashOfConfig":"11"},{"size":118,"mtime":1610760970624,"results":"14","hashOfConfig":"11"},{"size":2728,"mtime":1611928446410,"results":"15","hashOfConfig":"11"},{"size":1279,"mtime":1611927354407,"results":"16","hashOfConfig":"11"},{"size":743,"mtime":1611928446410,"results":"17","hashOfConfig":"11"},{"size":179,"mtime":1611927354404,"results":"18","hashOfConfig":"11"},{"size":3718,"mtime":1612143705738,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"nmrplc",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/jyj/Documents/Projects/showmemydog/src/index.tsx",[],[],"/Users/jyj/Documents/Projects/showmemydog/src/Routes.tsx",[],"/Users/jyj/Documents/Projects/showmemydog/src/Pages/Quiz/index.tsx",[],"/Users/jyj/Documents/Projects/showmemydog/src/Pages/Result/index.tsx",[],"/Users/jyj/Documents/Projects/showmemydog/src/Pages/Main/index.tsx",[],"/Users/jyj/Documents/Projects/showmemydog/src/Styles/globalStyles.ts",[],"/Users/jyj/Documents/Projects/showmemydog/src/Styles/theme.ts",[],"/Users/jyj/Documents/Projects/showmemydog/src/API.ts",[],"/Users/jyj/Documents/Projects/showmemydog/src/Pages/Quiz/cards.tsx",[]]
