+++ src/Pages/Quiz/types.ts
